Fix bug in closing audio tracks
This fixes a bug where an audio track would be closed twice by mistake,
leading to uncaught exceptions in the console. The issue was caused by
the track being closed inside of the cleanup function which would run
first when tracksToPull would change, and then a second time after
pulledTrackRecord would change.